Moved interceptors into common/interceptors package
