+++ src/main/java/io/temporal/common/interceptors/WorkflowCallsInterceptor.java
@@ -17,12 +17,16 @@
  *  permissions and limitations under the License.
  */
 
-package io.temporal.workflow;
+package io.temporal.common.interceptors;
 
 import io.temporal.activity.ActivityOptions;
 import io.temporal.activity.LocalActivityOptions;
 import io.temporal.proto.common.WorkflowExecution;
+import io.temporal.workflow.ChildWorkflowOptions;
+import io.temporal.workflow.ContinueAsNewOptions;
+import io.temporal.workflow.Functions;
 import io.temporal.workflow.Functions.Func;
+import io.temporal.workflow.Promise;
 import java.lang.reflect.Type;
 import java.time.Duration;
 import java.util.Map;
