+++ src/main/java/io/temporal/internal/sync/ExternalWorkflowInvocationHandler.java
@@ -21,12 +21,12 @@ package io.temporal.internal.sync;
 
 import static io.temporal.internal.common.InternalUtils.getValueOrDefault;
 
+import io.temporal.common.interceptors.WorkflowCallsInterceptor;
 import io.temporal.internal.common.InternalUtils;
 import io.temporal.proto.common.WorkflowExecution;
 import io.temporal.workflow.ExternalWorkflowStub;
 import io.temporal.workflow.QueryMethod;
 import io.temporal.workflow.SignalMethod;
-import io.temporal.workflow.WorkflowCallsInterceptor;
 import io.temporal.workflow.WorkflowMethod;
 import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.Method;
