+++ src/main/java/io/temporal/internal/sync/LocalActivityInvocationHandler.java
@@ -22,8 +22,8 @@ package io.temporal.internal.sync;
 import io.temporal.activity.ActivityMethod;
 import io.temporal.activity.LocalActivityOptions;
 import io.temporal.common.MethodRetry;
+import io.temporal.common.interceptors.WorkflowCallsInterceptor;
 import io.temporal.workflow.ActivityStub;
-import io.temporal.workflow.WorkflowCallsInterceptor;
 import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.Method;
 import java.util.function.Function;
