+++ src/main/java/io/temporal/internal/sync/POJOWorkflowImplementationFactory.java
@@ -26,6 +26,10 @@ import com.google.common.reflect.TypeToken;
 import io.temporal.common.context.ContextPropagator;
 import io.temporal.common.converter.DataConverter;
 import io.temporal.common.converter.DataConverterException;
+import io.temporal.common.interceptors.WorkflowCallsInterceptor;
+import io.temporal.common.interceptors.WorkflowInterceptor;
+import io.temporal.common.interceptors.WorkflowInvocationInterceptor;
+import io.temporal.common.interceptors.WorkflowInvoker;
 import io.temporal.internal.common.CheckedExceptionWrapper;
 import io.temporal.internal.common.InternalUtils;
 import io.temporal.internal.metrics.MetricsType;
@@ -41,11 +45,7 @@ import io.temporal.workflow.Functions.Func;
 import io.temporal.workflow.QueryMethod;
 import io.temporal.workflow.SignalMethod;
 import io.temporal.workflow.Workflow;
-import io.temporal.workflow.WorkflowCallsInterceptor;
 import io.temporal.workflow.WorkflowInfo;
-import io.temporal.workflow.WorkflowInterceptor;
-import io.temporal.workflow.WorkflowInvocationInterceptor;
-import io.temporal.workflow.WorkflowInvoker;
 import io.temporal.workflow.WorkflowMethod;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
