+++ src/main/java/io/temporal/internal/sync/WorkflowInternal.java
@@ -26,6 +26,7 @@ import com.uber.m3.tally.Scope;
 import io.temporal.activity.ActivityOptions;
 import io.temporal.activity.LocalActivityOptions;
 import io.temporal.common.RetryOptions;
+import io.temporal.common.interceptors.WorkflowCallsInterceptor;
 import io.temporal.internal.common.CheckedExceptionWrapper;
 import io.temporal.internal.common.InternalUtils;
 import io.temporal.internal.logging.ReplayAwareLogger;
@@ -42,7 +43,6 @@ import io.temporal.workflow.Functions.Func;
 import io.temporal.workflow.Promise;
 import io.temporal.workflow.QueryMethod;
 import io.temporal.workflow.Workflow;
-import io.temporal.workflow.WorkflowCallsInterceptor;
 import io.temporal.workflow.WorkflowInfo;
 import io.temporal.workflow.WorkflowQueue;
 import java.lang.reflect.InvocationHandler;
