+++ src/test/java/io/temporal/workflow/MetricsTest.java
@@ -34,6 +34,11 @@ import io.temporal.client.WorkflowClient;
 import io.temporal.client.WorkflowClientOptions;
 import io.temporal.client.WorkflowOptions;
 import io.temporal.common.RetryOptions;
+import io.temporal.common.interceptors.BaseWorkflowInvoker;
+import io.temporal.common.interceptors.WorkflowCallsInterceptor;
+import io.temporal.common.interceptors.WorkflowInterceptor;
+import io.temporal.common.interceptors.WorkflowInvocationInterceptor;
+import io.temporal.common.interceptors.WorkflowInvoker;
 import io.temporal.internal.metrics.MetricsTag;
 import io.temporal.internal.metrics.MetricsType;
 import io.temporal.testing.TestEnvironmentOptions;
