+++ src/test/java/io/temporal/workflow/WorkflowTest.java
@@ -45,6 +45,11 @@ import io.temporal.common.CronSchedule;
 import io.temporal.common.MethodRetry;
 import io.temporal.common.RetryOptions;
 import io.temporal.common.converter.JsonDataConverter;
+import io.temporal.common.interceptors.BaseWorkflowInvoker;
+import io.temporal.common.interceptors.WorkflowCallsInterceptor;
+import io.temporal.common.interceptors.WorkflowInterceptor;
+import io.temporal.common.interceptors.WorkflowInvocationInterceptor;
+import io.temporal.common.interceptors.WorkflowInvoker;
 import io.temporal.internal.common.QueryResponse;
 import io.temporal.internal.common.WorkflowExecutionUtils;
 import io.temporal.internal.sync.DeterministicRunnerTest;
