+++ src/test/java/io/temporal/workflow/interceptors/SignalWorkflowCallsInterceptor.java
@@ -21,6 +21,7 @@ package io.temporal.workflow.interceptors;
 
 import io.temporal.activity.ActivityOptions;
 import io.temporal.activity.LocalActivityOptions;
+import io.temporal.common.interceptors.WorkflowCallsInterceptor;
 import io.temporal.proto.common.WorkflowExecution;
 import io.temporal.workflow.*;
 import java.lang.reflect.Type;
